(PUP-7486) useradd provider silently does not set password
If password-related parameters are set on User resources the client
should alarm when unable to enforce those parameters on the host. When
`ruby-libshadow` is unavailable the provider won't/can't set the
password-related parameters, but until now there was no indication in
the output on a client or server that the parameters were not applied.

A side affect of this warning is the `:manages_passwords` and
`:manages_password_age` features are always "on" for the `useradd` provider
in order to alarm during the application of the catalog.